techpack: data: remove from kernel tree again
Building rmnet-perf into the kernel results in CFI-Violations causing the kernel to reboot, because we are running on CFI enforcing for additional security.
Lets remove this driver from the tree for now.

Additioanlly
ls /sys/module | grep rmnet on stock kernel running on mobile data shows only
qmi_rmnet and rmnet_ipa as dlkms.

Despite google pushing the source for the rmnet extensions it is not used in the stock kernel.