iommu: arm-smmu: clear TBU halt request when halt times out
TBU halt request is considered timed out after some wait.
But the halt request is not cleared. This can result in halt request
succeeding subsequently and driver not resuming the TBU. This can cause
functional issue of TBU power collapse or TBU remaining halted.
Clear the TBU halt request in case of timeout.

Change-Id: I2225485cae8146917ea8cfdf3bf50742278ae628
Signed-off-by: Prakash Gupta <guptap@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>